simple_camel/requirements.txt
pycamel>=0.1.0
pyyaml>=6.0
loguru>=0.7.0
python-dotenv>=1.0.0
decorator>=5.1.1
fastapi>=0.95.0
uvicorn>=0.22.0
jinja2>=3.1.2
aiofiles>=0.8.0
watchdog>=3.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
imaplib2>=3.6
email-validator>=2.0.0
httpx>=0.24.0
websockets>=11.0.3
